feat(k8s-ui): extract shared ApplicationsView core for single + fleet variants
ApplicationsList re-rendered its whole chassis (facet rail, hero, search,
sort, fold, keyboard nav, table) inside one file. The Cloud fleet view needs
the same chrome over a per-cluster-cell row shape, so extract a variant-
discriminated core: SingleAppEntry | FleetAppEntry drive one ApplicationsView;
ApplicationsList stays a thin single-variant wrapper (unchanged public API).

Fleet variant adds a Cluster column + multi-env pivot cells + version-skew;
single variant render is unchanged. emptySlot lets a host inject a coverage-
aware empty state.
